+++ lib/Doctrine/Annotations/Assembler/Validator/Constraint/RequiredConstraint.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Doctrine\Annotations\Assembler\Validator\Constraint;
 
+use Doctrine\Annotations\Assembler\Validator\Constraint\Exception\MissingRequiredValue;
+
 final class RequiredConstraint implements Constraint
 {
     /**
