+++ tests/Doctrine/Tests/Annotations/Metadata/Constraint/CompositeConstraintTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Doctrine\Tests\Annotations\Metadata\Constraint;
 
 use Doctrine\Annotations\Assembler\Validator\Constraint\CompositeConstraint;
-use Doctrine\Annotations\Assembler\Validator\Constraint\ConstraintNotFulfilled;
+use Doctrine\Annotations\Assembler\Validator\Constraint\Exception\ConstraintNotFulfilled;
 use Doctrine\Annotations\Assembler\Validator\Constraint\RequiredConstraint;
 use Doctrine\Annotations\Assembler\Validator\Constraint\TypeConstraint;
 use Doctrine\Annotations\Metadata\Type\StringType;
