+++ tests/Doctrine/Tests/Annotations/Metadata/Constraint/EnumConstraintTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Doctrine\Tests\Annotations\Metadata\Constraint;
 
 use Doctrine\Annotations\Assembler\Validator\Constraint\EnumConstraint;
-use Doctrine\Annotations\Assembler\Validator\Constraint\InvalidValue;
+use Doctrine\Annotations\Assembler\Validator\Constraint\Exception\InvalidValue;
 use PHPUnit\Framework\TestCase;
 
 final class EnumConstraintTest extends TestCase
